requirements.txt
pandas
fastapi
uvicorn
streamlit
joblib
numpy
xgboost
scikit-learn
PyMuPDF
matplotlib
plotly
pdf2image
python-docx
ipython

altair
requests
Pillow
pytesseract

python-multipart
imbalanced-learn